Setup utils script with say and fail functions
This script will be sourced from the main build script and aids in making both the script and its output better readable. The say function prints a given message to the stdout (or stderr) based on the selected formatting mode, and the fail function calls say with error mode and exits with 1.